Fix the ui language override helper method
